test(canon): split generated codegen tsconfig coverage
